Moved KlinkApiController tests to DocumentPreviewControllerTest
