+++ tests/KlinkApiControllerTest.php
@@ -19,207 +19,5 @@ class KlinkApiControllerTest extends BrowserKitTestCase
      *
      * @return void
      */
-    public function testDocumentShowForDocumentInProject()
-    {
-        $this->withKlinkAdapterFake();
-
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $user_accessing_the_document = $this->createUser(Capability::$PARTNER);
-        
-        $service = app('Klink\DmsDocuments\DocumentsService');
-
-        $document = $this->createDocument($user);
-
-        $project1 = $this->createProject(['user_id' => $user->id]);
-        $project1->users()->attach($user_accessing_the_document->id);
-
-        $project1_child1 = $this->createProjectCollection($user, $project1);
-        $service->addDocumentToGroup($user, $document, $project1_child1);
-        
-        $url = route('klink_api', ['id' => $document->local_document_id, 'action' => 'document']);
-
-        $this->actingAs($user_accessing_the_document);
-        
-        $this->visit($url);
-        
-        $this->assertResponseOk();
-
-        $this->assertViewName('documents.preview');
-    }
-
-    public function testDocumentShowForSharedDocument()
-    {
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $user_accessing_the_document = $this->createUser(Capability::$PARTNER);
-
-        $document = $this->createDocument($user);
-
-        $document->shares()->create([
-            'user_id' => $user->id,
-            'sharedwith_id' => $user_accessing_the_document->id, //the id
-            'sharedwith_type' => get_class($user_accessing_the_document), //the class
-            'token' => hash('sha512', '$token_content'),
-        ]);
-
-        
-        $url = route('klink_api', ['id' => $document->local_document_id, 'action' => 'document']);
-
-        $this->actingAs($user_accessing_the_document);
-        
-        $this->visit($url);
-
-        $this->assertViewName('documents.preview');
-    }
-
-    public function testDocumentShowForPublicDocument()
-    {
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $user_accessing_the_document = $this->createUser(Capability::$PARTNER);
-
-        $document = $this->createDocument($user, 'public');
-
-        Publication::unguard(); // as fields are not mass assignable
-        
-        $document->publications()->create([
-            'published_at' => Carbon::now()
-        ]);
-        
-        $url = route('klink_api', ['id' => $document->local_document_id, 'action' => 'document']);
-
-        $this->actingAs($user_accessing_the_document);
-        
-        $this->visit($url);
-
-        $this->assertViewName('documents.preview');
-    }
-
-    public function testDocumentShowForPublicDocumentWithNoLogin()
-    {
-        $this->withKlinkAdapterFake();
-
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $user_accessing_the_document = $this->createUser(Capability::$PARTNER);
-
-        $service = app('Klink\DmsDocuments\DocumentsService');
-
-        $document = $this->createDocument($user, 'public');
-
-        Publication::unguard(); // as fields are not mass assignable
-        
-        $document->publications()->create([
-            'published_at' => Carbon::now()
-        ]);
-
-        $project1 = $this->createProject(['user_id' => $user->id]);
-        $project1->users()->attach($user_accessing_the_document->id);
-
-        $project1_child1 = $this->createProjectCollection($user, $project1);
-        $service->addDocumentToGroup($user, $document, $project1_child1);
-
-        $url = route('klink_api', ['id' => $document->local_document_id, 'action' => 'document']);
-        
-        $this->visit($url);
-
-        $this->assertViewName('documents.preview');
-    }
     
-    public function testDocumentShowForDocumentNotSharedNorInProject()
-    {
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $user_accessing_the_document = $this->createUser(Capability::$PARTNER);
-
-        $document = $this->createDocument($user);
-
-        
-        $url = route('klink_api', ['id' => $document->local_document_id, 'action' => 'document']);
-
-        $this->actingAs($user_accessing_the_document);
-        
-        $this->visit($url);
-
-        $this->assertViewName('errors.403');
-    }
-
-    public function testDocumentShowForPrivateDocumentNotInCollectionOrShared()
-    {
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $user_accessing_the_document = $this->createUser(Capability::$PARTNER);
-
-        $document = $this->createDocument($user, 'private');
-
-        
-        $url = route('klink_api', ['id' => $document->local_document_id, 'action' => 'document']);
-
-        $this->actingAs($user);
-        
-        $this->visit($url);
-
-        $this->assertViewName('documents.preview');
-    }
-    
-    public function testDocumentShowForDocumentInProjectWithOwnerDisabled()
-    {
-        $this->withKlinkAdapterFake();
-
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $user_accessing_the_document = $this->createUser(Capability::$PARTNER);
-        
-        $service = app('Klink\DmsDocuments\DocumentsService');
-
-        $document = $this->createDocument($user);
-
-        $project1 = $this->createProject(['user_id' => $user->id]);
-        $project1->users()->attach($user_accessing_the_document->id);
-
-        $project1_child1 = $this->createProjectCollection($user, $project1);
-        $service->addDocumentToGroup($user, $document, $project1_child1);
-        
-        $url = route('klink_api', ['id' => $document->local_document_id, 'action' => 'document']);
-
-        $user->delete();
-
-        $this->actingAs($user_accessing_the_document);
-        
-        $this->visit($url);
-        
-        $this->assertResponseOk();
-
-        $this->assertViewName('documents.preview');
-    }
-
-    public function testDocumentShowForPrivateDocumentNotInCollectionOrSharedViaOtherUser()
-    {
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $user_accessing_the_document = $this->createUser(Capability::$PARTNER);
-
-        $document = $this->createDocument($user, 'private');
-
-        
-        $url = route('klink_api', ['id' => $document->local_document_id, 'action' => 'document']);
-
-        $this->actingAs($user_accessing_the_document);
-        
-        $this->visit($url);
-
-        $this->assertViewName('errors.403');
-    }
-
-    public function testDocumentNotFoundIfFileIsTrashed()
-    {
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        
-        $document = $this->createDocument($user, 'private');
-        
-        $document->file_id = null;
-        $document->save();
-        
-        $url = route('klink_api', ['id' => $document->local_document_id, 'action' => 'document']);
-        
-        $this->actingAs($user);
-        
-        $this->disableExceptionHandling();
-        $this->expectException(\Symfony\Component\HttpKernel\Exception\NotFoundHttpException::class);
-        
-        $this->visit($url);
-    }
 }
